+++ api/pom.xml
@@ -12,10 +12,6 @@
 	<name>EMR API Module API</name>
 	<description>API project for EMRAPI</description>
 
-    <properties>
-        <openMRSVersion>1.9.9</openMRSVersion>
-    </properties>
-
     <dependencies>
 
         <dependency>
