+++ api/src/main/java/org/openmrs/module/emrapi/encounter/mapper/OrderMapper1_12.java
@@ -33,7 +33,6 @@ import java.util.Comparator;
 import java.util.List;
 
 @Component(value = "orderMapper")
-@OpenmrsProfile(openmrsPlatformVersion = "1.12.0 - 2.*")
 public class OrderMapper1_12 implements OrderMapper {
 
     private final ConceptMapper conceptMapper = new ConceptMapper();
